requirements.txt
qaseio>=3.3.0
certifi
requests
mysql-connector-python